Reduce exception-related bloat in libnanobind
The many custom exception definitions to mirror Python variants (e.g.
``nb::index_error`` -> ``IndexError``) added many symbols (virtual
tables, RTTI type information) that needed to be exported globally for
these to work correctly in a multi-module setting.

This commit refactors the exception class hierarchy so that most of
these are new represented by the generic ``nb::builtin_exception``.

For example, ``nb::index_error`` is no longer a type: it was replaced by
a function returning a suitable ``nb::builtin_exception`` instance. This
should not be a breaking change, as the documentation clearly advertised
that users should only throw, but never catch those types of exceptions.

``nb::cast_error`` has turned into an alias for the ``std::bad_cast``.

The commit also bundles a tiny tweak: a ``std::vector`` for the
exception translator list was replaced by a hand-written singly linked
list, which reduced size further.

+++ include/nanobind/stl/chrono.h
@@ -206,7 +206,7 @@ public:
                              localtime.tm_hour,
                              localtime.tm_min,
                              localtime.tm_sec,
-                             us.count());
+                             (int) us.count());
     }
     NB_TYPE_CASTER(type, const_name("datetime.datetime"));
 };
